blogs

+++ src/main/java/com/driver/services/BlogService.java
@@ -41,13 +41,15 @@ public class BlogService {
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
-        Blog blog =  blogRepository1.findById(blogId).get();
-        User user = blog.getUser();
-        List<Blog> blogList = user.getBlogList();
-        blogList.remove(blog);
-        user.setBlogList(blogList);
+        if (blogRepository1.existsById(blogId)){
+            Blog blog =  blogRepository1.findById(blogId).get();
+            User user = blog.getUser();
+            List<Blog> blogList = user.getBlogList();
+            blogList.remove(blog);
+            user.setBlogList(blogList);
 
-        userRepository1.save(user);
-        blogRepository1.deleteById(blogId);
+            userRepository1.save(user);
+            blogRepository1.deleteById(blogId);
+        }
     }
 }
